# docs/swagger.yaml
definitions:
...
  courses.CreateCourseRequest:
    properties:
      course_credit:
        maximum: 12
        minimum: 1
        type: integer
      course_day:
        allOf:
        - $ref: '#/definitions/models.Day'
        enum:
        - monday
        - tuesday
        - wednesday
        - thursday
        - friday
        - saturday
        - sunday
      course_department:
        type: string
      course_end_shift:
        maximum: 10
        minimum: 1
        type: integer
      course_fullname:
        maxLength: 200
        minLength: 5
        type: string
      course_name:
        maxLength: 100
        minLength: 3
        type: string
      course_room:
        maxLength: 50
        minLength: 2
        type: string
      course_semester:
        enum:
        - 1
        - 2
        - 3
        type: integer
      course_start_shift:
        maximum: 10
        minimum: 1
        type: integer
      course_teacher_id:
        description: CourseCode       string     `json:"course_code"`
        type: string
      course_year:
        maximum: 2100
        minimum: 2000
        type: integer
      current_enroll:
        maximum: 1000
        minimum: 0
        type: integer
      max_enroll:
        maximum: 1000
        minimum: 1
        type: integer
    required:
    - course_credit
    - course_day
    - course_department
    - course_end_shift
    - course_fullname
    - course_name
    - course_room
    - course_semester
    - course_start_shift
    - course_teacher_id
    - course_year
    - max_enroll
    type: object
  courses.EditCourseRequest:
    properties:
      course_credit:
        type: integer
      course_day:
        $ref: '#/definitions/models.Day'
      course_department:
        type: string
      course_end_shift:
        type: integer
      course_fullname:
        type: string
      course_id:
        type: string
      course_name:
        type: string
      course_room:
        type: string
      course_semester:
        type: integer
      course_start_shift:
        type: integer
      course_teacher_id:
        type: string
      course_year:
        type: integer
      current_enroll:
        type: integer
      max_enroll:
        type: integer
    required:
    - course_id
    type: object
...
  models.Day:
    enum:
    - monday
    - tuesday
    - wednesday
    - thursday
    - friday
    - saturday
    - sunday
    type: string
    x-enum-varnames:
    - DayMonday
    - DayTuesday
    - DayWednesday
    - DayThursday
    - DayFriday
    - DaySaturday
    - DaySunday
...
  models.Response:
    properties:
      data: {}
      message:
        type: string
      status:
        type: boolean
    type: object
...
  schemas.CreateCourseResponse:
    properties:
      course_credit:
        type: integer
      course_day:
        $ref: '#/definitions/models.Day'
      course_department:
        type: string
      course_end_shift:
        type: integer
      course_fullname:
        type: string
      course_name:
        type: string
      course_room:
        type: string
      course_semester:
        type: integer
      course_start_shift:
        type: integer
      course_teacher_id:
        type: string
      course_teacher_name:
        type: string
      course_year:
        type: integer
      current_enroll:
        type: integer
      id:
        type: string
      max_enroll:
        type: integer
    type: object
...
info:
  contact: {}
paths:
  /course/confirm:
    put:
      consumes:
      - application/json
      description: Confirm course
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Course ID
        in: query
        name: course_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Confirm Course
      tags:
      - Courses
  /course/create:
    post:
      consumes:
      - application/json
      description: Create course
      parameters:
      - description: CreateCourseRequest
        in: body
        name: CreateCourseRequest
        required: true
        schema:
          $ref: '#/definitions/courses.CreateCourseRequest'
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Create course
      tags:
      - Course
  /course/delete:
    delete:
      consumes:
      - application/json
      description: Delete course
      parameters:
      - description: Course ID
        in: body
        name: course_id
        required: true
        schema:
          type: string
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Delete course
      tags:
      - Courses
  /course/edit:
    put:
      consumes:
      - application/json
      description: Edit course
      parameters:
      - description: EditCourseRequest
        in: body
        name: EditCourseRequest
        required: true
        schema:
          $ref: '#/definitions/courses.EditCourseRequest'
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Edit course
      tags:
      - Course
  /course/list:
    get:
      consumes:
      - application/json
      description: List all courses
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.CreateCourseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: List Courses
      tags:
      - Courses
...
swagger: "2.0"